2.2
This was written over a year ago, but it was the beginnings of a v3.0
for the script. Many things are not working, and sorting through this
code is proving difficult. However, if I am able to sort through it and
comment it out it is OOP ready and much more organized that the previous
v2.1 which will make adding new features such as switching the Left and
Right analog stick functions very easy to do in the config file.