2.2
Added comments to constructor.
Began updating variable names to conform to a naming scheme. Tedious
process to do in notepad++
Made the stick threshold a vector2

Updated some config names to match changes to variable names. More
changes are needed.

Added a System.ahk to add global useful classes. Started with Vector2

Created a workspace for notepad++. Not very intuitive, but I got it
working. Helps with organizing, just not quick to reopen it.