Set in Config
Folder restructure to allow for a default config.ini and preferences.ini
to be stored. This will allow the user to go back to the default values
if they mess anything up once I've programmed it in.

Added code in the AHK and WinForm to set a preferences location. This
allows for different files to be used similar to a profile type of
setup.