Added sensitivity values, targeting delay, and fixed bugs found when playing PoE
Added values in the profile to allow for modifying the sensitivity of the "Freedom" and "FreeTargeting" modes. Each axis is configurable separately.

Added a value in the profile to allow for delaying the time between moving the mouse to the targeting reticule position and firing off a hotkey. This change allows skills to work properly in PoE.

Fixed issue where the "ForceMoveKey" was getting set to the literal "Space" rather than back to the default value after using a hotkey which ignored the targeting reticule. This bug wasn't picked up by anyone since in Diablo III the default force move is "Space".

Fixed and issue where while in "Freedom" mode, moving the analog stick pressed hotkeys.