Clarity on .ini comments, enable 'Cursor' or 'FreeTarget' mode on startup, Allowed Modifiers on special actions, Allowed all special actions on press or hold
Clarified a bunch of the comments in the default .ini profile to better explain each feature and value.

Added a value in the .ini to enable 'Cursor' or 'FreeTarget' mode on startup.

Fixed bugs with special actioins. Special actions can now properly contain modifiers and can now properly be activated on press or hold.

Lots of minor code cleanup.