Wrapping in Inventory Mode, and tooltip support
Added code to wrap the inventory grid around when reaching certain values to help navigate it easier.

Added tooltip support. A bunch of this should be converted to classes. I need to look up this syntax for autohotkey and start converting a bunch of the arrays I have to this since the arrays are just confusing to use. This is of course assuming that the class system in ahk is actually good.