Debug class up and running
Added a delegate class to help with adding text to the debug tooltip easier. That way showing variables or other text every frame as a sort of watcher window is super easy now. I'd like to add a text log(just putting the log into the file instead of only the tooltip), but adding actual functionality for all of the controller buttons should be higher on the list.