Debug Optimization, Inventory HoldToMove
Optimized the debug tooltip draw to not include the log if the log hasn't changed since the last update.

When in inventory mode, you can now hold the dpad button down if "Hold_To_Move" is set to true in the ini. this will allow the user to hold down a direction to move in the inventory grid.