Image Overlay, Bug Fixes
Added functionality for the Image Overlay. Looks really nice. Quite happy with it, and seems to have ~0 impact on performance as well.

Fixed a bunch of bugs related to the movement stick and keys which are set to be of type movement. Everything feels really solid now and if more issues come up it should be pretty simple to fix them.