Fixes for LoL
tried out the script in LoL. It works, but movement is nearly impossible in the river. I'm thinking about allowing for the movement stick to not be fixed at a certain radius, and instead move similarly to the target stick. It seems like the only really way to fix this issue. Setting a really high custom deadzone on the top end and allowing free movement may be the only way to play.