FPS, GUI Debug, DeltaTime sticks
Added an FPS class, although its use if very limited. For some reason, attempting to sleep at any point just sleeps for 16+ ms no matter what. Not FPS below 60 can be achieved using the class.

The debug info is in the process of being moved to a GUI draw which is much faster than tooltips.

The debug log now outputs to a text file as well.