Debug Resize
Resized the debug log gui as it gets longer. this is pretty expensive, so i'm looking into what i can do to speed this up. not a huge deal though honeslty since this only happens when debug is pulled up. Performance is still great with it off.

Fixed a bug with ellipse drawing.

need to remove singleton on Graphics so that I can add a callback for active window changing. Need to also add an event listener class to clean that process up.