The why am I doing this commit
Adding a new control scheme. WASD movement. Yes. WASD. It's not going too badly considering, but every line of code is a new hurdle. Need to post some questions to the forums to see if I can get some help. Performance is the main problem right now and it has nothing to do with my script but rather how Autohotkey handles threads. An empty loop taxes the callback function for the mouse...why?