Better League compatibility and a small graphics optimization
Updated the way that m_RepeatForceMove affects movement to actually work properly in League.

Stopped using Gui to draw the image at a position and instead moved the Gdip window. It's a huge improvement when using the debug overlay, but will not be particularly noticable outside. Still a very positive change considering a new feature I plan on adding which will take advantage of this optimization.