Target Triangle and Inventory Offset
Added a targeting triangle option to help show where the user is currently targeting if needed.

Added offset values for the inventory. This is pixel based and does not scale with resolution to assit with ease of use.

Added the relevant values to the default ini's for the new options.

Added the proper logic for updating ini's to the new version in the C# config application.